Add the loading indicator to all pages with loading state
