+++ client/src/pages/launches.js
@@ -1,13 +1,90 @@
 import React, { Fragment } from 'react';
+import { Query } from 'react-apollo';
+import gql from 'graphql-tag';
 
+import LaunchTile from '../components/launch-tile';
 import Header from '../components/header';
-import LaunchesList from '../containers/launches-list';
+import Button from '../components/button';
+import Loading from '../components/loading';
+
+export const LAUNCH_TILE_DATA = gql`
+  fragment LaunchTile on Launch {
+    id
+    isBooked
+    rocket {
+      id
+      name
+    }
+    mission {
+      name
+      missionPatch
+    }
+  }
+`;
+
+const GET_LAUNCHES = gql`
+  query launchList($after: String) {
+    isLoggedIn @client
+    launches(after: $after) {
+      cursor
+      hasMore
+      launches {
+        isInCart @client
+        ...LaunchTile
+      }
+    }
+  }
+  ${LAUNCH_TILE_DATA}
+`;
 
 export default function Launches() {
   return (
-    <Fragment>
-      <Header />
-      <LaunchesList />
-    </Fragment>
-  )
-}
+    <Query query={GET_LAUNCHES}>
+      {({ data, loading, error, fetchMore }) => {
+        if (loading) return <Loading />;
+        if (error) return <p>ERROR</p>;
+
+        return (
+          <Fragment>
+            <Header />
+            {data.launches &&
+              data.launches.launches &&
+              data.launches.launches.map((l, index) => (
+                <LaunchTile
+                  key={l.id}
+                  launch={l}
+                  isLoggedIn={data.isLoggedIn}
+                />
+              ))}
+            {data.launches && data.launches.hasMore ? (
+              <Button
+                onClick={() =>
+                  fetchMore({
+                    variables: {
+                      after: data.launches.cursor,
+                    },
+                    updateQuery: (prev, { fetchMoreResult, ...rest }) => {
+                      if (!fetchMoreResult) return prev;
+                      return {
+                        ...fetchMoreResult,
+                        launches: {
+                          ...fetchMoreResult.launches,
+                          launches: [
+                            ...prev.launches.launches,
+                            ...fetchMoreResult.launches.launches,
+                          ],
+                        },
+                      };
+                    },
+                  })
+                }
+              >
+                Load More
+            </Button>
+            ) : null}
+          </Fragment>
+        );
+      }}
+    </Query>
+  );
+};
